requirements.txt
flask
waitress
pyyaml
Flask-Login
pycryptodome
